Move stb files to subfolder
