+++ code/components/jomjol_image_proc/CImageBasis.h
@@ -11,9 +11,9 @@
 
 #include <math.h>
 
-#include "stb_image.h"
-#include "stb_image_write.h"
-#include "stb_image_resize.h"
+#include "stb/stb_image.h"
+#include "stb/stb_image_write.h"
+#include "stb/stb_image_resize.h"
 
 #include "esp_heap_caps.h"
