+++ code/components/jomjol_image_proc/make_stb.cpp
@@ -10,10 +10,10 @@
 #define STBI_FREE(p)              free_psram_heap("STBI", p)
 
 #define STB_IMAGE_IMPLEMENTATION
-#include "stb_image.h"
+#include "stb/stb_image.h"
 
 #define STB_IMAGE_WRITE_IMPLEMENTATION
-#include "stb_image_write.h"
+#include "stb/stb_image_write.h"
 
 #define STB_IMAGE_RESIZE_IMPLEMENTATION
-#include "stb_image_resize.h"
+#include "stb/stb_image_resize.h"
